pyproject.toml
[project]
name = "jobseeking-agent"
version = "0.1.0"
description = "Personal job application agent"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.0.0",
    "sqlmodel>=0.0.22",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "apscheduler>=3.10.0",
    "playwright>=1.40.0",
    "textual>=0.61.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "fastapi>=0.134.0",
    "uvicorn[standard]>=0.41.0",
    "python-multipart>=0.0.22",
]

[project.optional-dependencies]
build = ["pyinstaller>=6.5.0"]

[project.scripts]
scout = "jobseeking_agent.cli:scout"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jobseeking_agent"]
